BLD: renaming all instances of rhapsody to mmvec
