+++ CHANGELOG.md
@@ -1,19 +1,19 @@
-# rhapsody changelog
+# mmvec changelog
 
 ## Version 0.4.0 (2019-07-22)
 
 # Enhancements
- - Simpler standalone CLI interface - now all outputs have named rows and columns. [#61](https://github.com/biocore/rhapsody/pull/61)
+ - Simpler standalone CLI interface - now all outputs have named rows and columns. [#61](https://github.com/biocore/mmvec/pull/61)
 
 # Bug fixes
- - The ranks file is no longer empty. [#61](https://github.com/biocore/rhapsody/pull/61)
+ - The ranks file is no longer empty. [#61](https://github.com/biocore/mmvec/pull/61)
 
 ## Version 0.3.0 (2019-06-20)
 
 Initial beta release.
 
 # Bug fixes
- - Biplots are now being properly centered in the qiime2 interface [#58](https://github.com/biocore/rhapsody/pull/58)
+ - Biplots are now being properly centered in the qiime2 interface [#58](https://github.com/biocore/mmvec/pull/58)
 
 
 ## Version 0.2.0 (2019-04-22)
