+++ Makefile
@@ -1,6 +1,6 @@
 .DEFAULT_GOAL := help
 
-TEST_COMMAND = nosetests --with-doctest
+TEST_COMMAND = nosetests
 help:
 	@echo 'Use "make test" to run all the unit tests and docstring tests.'
 	@echo 'Use "make pep8" to validate PEP8 compliance.'
@@ -9,7 +9,7 @@ help:
 test:
 	$(TEST_COMMAND)
 pep8:
-	pycodestyle rhapsody setup.py
-	flake8 rhapsody setup.py
+	pycodestyle mmvec setup.py
+	flake8 mmvec setup.py
 
 all: pep8 test
