+++ mmvec/util.py
