+++ setup.py
@@ -34,12 +34,12 @@ with open('README.md') as f:
 # version parsing from __init__ pulled from Flask's setup.py
 # https://github.com/mitsuhiko/flask/blob/master/setup.py
 _version_re = re.compile(r'__version__\s+=\s+(.*)')
-with open('rhapsody/__init__.py', 'rb') as f:
+with open('mmvec/__init__.py', 'rb') as f:
     hit = _version_re.search(f.read().decode('utf-8')).group(1)
     version = str(ast.literal_eval(hit))
 
 
-setup(name='rhapsody',
+setup(name='mmvec',
       version=version,
       license='BSD-3-Clause',
       description=description,
@@ -50,7 +50,7 @@ setup(name='rhapsody',
       maintainer="gneiss development team",
       maintainer_email="jamietmorton@gmail.com",
       packages=find_packages(),
-      scripts=glob('scripts/rhapsody'),
+      scripts=glob('scripts/mmvec'),
       install_requires=[
           'biom-format',
           'numpy >= 1.9.2',
@@ -63,7 +63,7 @@ setup(name='rhapsody',
       ],
       classifiers=classifiers,
       entry_points={
-          'qiime2.plugins': ['q2-rhapsody=rhapsody.q2.plugin_setup:plugin']
+          'qiime2.plugins': ['q2-mmvec=mmvec.q2.plugin_setup:plugin']
       },
       package_data={},
       zip_safe=False)
